+++ client/proxgui.cpp
@@ -15,7 +15,7 @@
 static ProxGuiQT *gui = NULL;
 static WorkerThread *main_loop_thread = NULL;
 
-WorkerThread::WorkerThread(char *script_cmds_file, bool usb_present) : script_cmds_file(script_cmds_file), usb_present(usb_present)
+WorkerThread::WorkerThread(char *script_cmds_file, char *script_cmd, bool usb_present) : script_cmds_file(script_cmds_file), script_cmd(script_cmd), usb_present(usb_present)
 {
 }
 
@@ -24,7 +24,7 @@ WorkerThread::~WorkerThread()
 }
 
 void WorkerThread::run() {
-	main_loop(script_cmds_file, usb_present);
+	main_loop(script_cmds_file, script_cmd, usb_present);
 }
 
 extern "C" void ShowGraphWindow(void)
@@ -56,11 +56,10 @@ extern "C" void MainGraphics(void)
 	if (!gui)
 		return;
 
-	main_loop_thread->start();
 	gui->MainLoop();
 }
 
-extern "C" void InitGraphics(int argc, char **argv, char *script_cmds_file, bool usb_present)
+extern "C" void InitGraphics(int argc, char **argv, char *script_cmds_file, char *script_cmd, bool usb_present)
 {
 #ifdef Q_WS_X11
 	bool useGUI = getenv("DISPLAY") != 0;
@@ -70,18 +69,15 @@ extern "C" void InitGraphics(int argc, char **argv, char *script_cmds_file, bool
 	if (!useGUI)
 		return;
 
-	gui = new ProxGuiQT(argc, argv);
-	main_loop_thread = new WorkerThread(script_cmds_file, usb_present);
-	QObject::connect(main_loop_thread, SIGNAL(finished()), main_loop_thread, SLOT(deleteLater()));
-	QObject::connect(main_loop_thread, SIGNAL(finished()), gui, SLOT(_Exit()));
+	main_loop_thread = new WorkerThread(script_cmds_file, script_cmd, usb_present);
+	gui = new ProxGuiQT(argc, argv, main_loop_thread);
 }
-
 
 extern "C" void ExitGraphics(void)
 {
-  if (!gui)
-    return;
+	if (!gui)
+		return;
 
-  gui->Exit();
-  gui = NULL;
+	gui->Exit();
+	gui = NULL;
 }