+++ client/proxmark3.c
@@ -18,14 +18,21 @@
 #include <readline/history.h>
 
 #include "proxmark3.h"
+#include "util_posix.h"
 #include "proxgui.h"
 #include "cmdmain.h"
 #include "uart.h"
 #include "ui.h"
+#include "util.h"
 #include "cmdparser.h"
 #include "cmdhw.h"
 #include "whereami.h"
 
+#ifdef _WIN32
+#define SERIAL_PORT_H	"com3"
+#else
+#define SERIAL_PORT_H	"/dev/ttyACM0"
+#endif
 
 // a global mutex to prevent interlaced printing from different threads
 pthread_mutex_t print_lock;
@@ -89,11 +96,13 @@ static void *uart_receiver(void *targ) {
 }
 
 
-void main_loop(char *script_cmds_file, bool usb_present) {
+void main_loop(char *script_cmds_file, char *script_cmd, bool usb_present) {
 	struct receiver_arg rarg;
 	char *cmd = NULL;
 	pthread_t reader_thread;
-
+	bool execCommand = (script_cmd != NULL);
+	bool stdinOnPipe = !isatty(STDIN_FILENO);
+	
 	if (usb_present) {
 		rarg.run = 1;
 		pthread_create(&reader_thread, NULL, &uart_receiver, &rarg);
@@ -101,46 +110,68 @@ void main_loop(char *script_cmds_file, bool usb_present) {
 		CmdVersion(NULL);
 	}
 
+	// file with script
 	FILE *script_file = NULL;
-	char script_cmd_buf[256];  // iceman, needs lua script the same file_path_buffer as the rest
+	char script_cmd_buf[256] = {0};  // iceman, needs lua script the same file_path_buffer as the rest
 
 	if (script_cmds_file) {
 		script_file = fopen(script_cmds_file, "r");
 		if (script_file) {
-			printf("using 'scripting' commands file %s\n", script_cmds_file);
+			printf("executing commands from file: %s\n", script_cmds_file);
 		}
 	}
-
+	
 	read_history(".history");
 
 	while(1)  {
-
 		// If there is a script file
 		if (script_file)
 		{
+			memset(script_cmd_buf, 0, sizeof(script_cmd_buf));
 			if (!fgets(script_cmd_buf, sizeof(script_cmd_buf), script_file)) {
 				fclose(script_file);
 				script_file = NULL;
 			} else {
-				char *nl;
-				nl = strrchr(script_cmd_buf, '\r');
-				if (nl) *nl = '\0';
-				
-				nl = strrchr(script_cmd_buf, '\n');
-				if (nl) *nl = '\0';
+				strcleanrn(script_cmd_buf, sizeof(script_cmd_buf));
 
-				if ((cmd = (char*) malloc(strlen(script_cmd_buf) + 1)) != NULL) {
-					memset(cmd, 0, strlen(script_cmd_buf));
-					strcpy(cmd, script_cmd_buf);
-					printf("%s\n", cmd);
+				if ((cmd = strmcopy(script_cmd_buf)) != NULL) {
+					printf(PROXPROMPT"%s\n", cmd);
+				}
+			}
+		} else {
+			// If there is a script command
+			if (execCommand){
+				if ((cmd = strmcopy(script_cmd)) != NULL) {
+					printf(PROXPROMPT"%s\n", cmd);
+				}
+
+				execCommand = false;
+			} else {
+				// exit after exec command
+				if (script_cmd)
+					break;
+
+				// if there is a pipe from stdin
+				if (stdinOnPipe) {
+					memset(script_cmd_buf, 0, sizeof(script_cmd_buf));
+					if (!fgets(script_cmd_buf, sizeof(script_cmd_buf), stdin)) {
+						printf("\nStdin end. Exit...\n");
+						break;
+					}
+					strcleanrn(script_cmd_buf, sizeof(script_cmd_buf));
+
+					if ((cmd = strmcopy(script_cmd_buf)) != NULL) {
+						printf(PROXPROMPT"%s\n", cmd);
+					}
+					
+				} else {		
+					// read command from command prompt
+					cmd = readline(PROXPROMPT);
 				}
 			}
 		}
 		
-		if (!script_file) {
-			cmd = readline(PROXPROMPT);
-		}
-		
+		// execute command
 		if (cmd) {
 
 			while(cmd[strlen(cmd) - 1] == ' ')
@@ -154,12 +185,13 @@ void main_loop(char *script_cmds_file, bool usb_present) {
 				}
 			}
 			free(cmd);
+			cmd = NULL;
 		} else {
 			printf("\n");
 			break;
 		}
 	}
-  
+
 	write_history(".history");
   
 	if (usb_present) {
@@ -171,7 +203,6 @@ void main_loop(char *script_cmds_file, bool usb_present) {
 		fclose(script_file);
 		script_file = NULL;
 	}
-
 }
 
 static void dumpAllHelp(int markdown)
@@ -212,36 +243,129 @@ static void set_my_executable_path(void)
 	}
 }
 
+static void show_help(bool showFullHelp, char *command_line){
+	printf("syntax: %s <port> [-h|-help|-m|-f|-flush|-w|-wait|-c|-command|-l|-lua] [cmd_script_file_name] [command][lua_script_name]\n", command_line);
+	printf("\tLinux example:'%s /dev/ttyACM0'\n", command_line);
+	printf("\tWindows example:'%s com3'\n\n", command_line);
+	
+	if (showFullHelp){
+		printf("help: <-h|-help> Dump all interactive command's help at once.\n");
+		printf("\t%s  -h\n\n", command_line);
+		printf("markdown: <-m> Dump all interactive help at once in markdown syntax\n");
+		printf("\t%s -m\n\n", command_line);
+		printf("flush: <-f|-flush> Output will be flushed after every print.\n");
+		printf("\t%s -f\n\n", command_line);
+		printf("wait: <-w|-wait> 20sec waiting the serial port to appear in the OS\n");
+		printf("\t%s "SERIAL_PORT_H" -w\n\n", command_line);
+		printf("script: A script file with one proxmark3 command per line.\n\n");
+		printf("command: <-c|-command> Execute one proxmark3 command.\n");
+		printf("\t%s "SERIAL_PORT_H" -c \"hf mf chk 1* ?\"\n", command_line);
+		printf("\t%s "SERIAL_PORT_H" -command \"hf mf nested 1 *\"\n\n", command_line);
+		printf("lua: <-l|-lua> Execute lua script.\n");
+		printf("\t%s "SERIAL_PORT_H" -l hf_read\n\n", command_line);
+	}
+}
 
 int main(int argc, char* argv[]) {
 	srand(time(0));
   
+	bool usb_present = false;
+	bool waitCOMPort = false;
+	bool executeCommand = false;
+	bool addLuaExec = false;
+	char *script_cmds_file = NULL;
+	char *script_cmd = NULL;
+  
 	if (argc < 2) {
-		printf("syntax: %s <port>\n\n",argv[0]);
-		printf("\tLinux example:'%s /dev/ttyACM0'\n\n", argv[0]);
-		printf("help:   %s -h\n\n", argv[0]);
-		printf("\tDump all interactive help at once\n");
-		printf("markdown:   %s -m\n\n", argv[0]);
-		printf("\tDump all interactive help at once in markdown syntax\n");
+		show_help(true, argv[0]);
 		return 1;
 	}
-	if (strcmp(argv[1], "-h") == 0) {
-		printf("syntax: %s <port>\n\n",argv[0]);
-		printf("\tLinux example:'%s /dev/ttyACM0'\n\n", argv[0]);
-		dumpAllHelp(0);
-		return 0;
-	}
-	if (strcmp(argv[1], "-m") == 0) {
-		dumpAllHelp(1);
-		return 0;
+
+	for (int i = 1; i < argc; i++) {
+		if (strcmp(argv[i], "-h") == 0 || strcmp(argv[i],"-help") == 0) {
+			show_help(false, argv[0]);
+			dumpAllHelp(0);
+			return 0;
+		}
+		
+		if (strcmp(argv[i], "-m") == 0) {
+			dumpAllHelp(1);
+			return 0;
+		}
+		
+		if(strcmp(argv[i],"-f") == 0 || strcmp(argv[i],"-flush") == 0){
+			printf("Output will be flushed after every print.\n");
+			flushAfterWrite = 1;
+		}
+		
+		if(strcmp(argv[i],"-w") == 0 || strcmp(argv[i],"-wait") == 0){
+			waitCOMPort = true;
+		}
+
+		if(strcmp(argv[i],"-c") == 0 || strcmp(argv[i],"-command") == 0){
+			executeCommand = true;
+		}
+
+		if(strcmp(argv[i],"-l") == 0 || strcmp(argv[i],"-lua") == 0){
+			executeCommand = true;
+			addLuaExec = true;
+		}
 	}
 
+	// If the user passed the filename of the 'script' to execute, get it from last parameter
+	if (argc > 2 && argv[argc - 1] && argv[argc - 1][0] != '-') {
+		if (executeCommand){
+			script_cmd = argv[argc - 1];
+			
+			while(script_cmd[strlen(script_cmd) - 1] == ' ')
+				script_cmd[strlen(script_cmd) - 1] = 0x00;
+			
+			if (strlen(script_cmd) == 0) {
+				script_cmd = NULL;
+			} else {
+				if (addLuaExec){
+					// add "script run " to command
+					char *ctmp = NULL;
+					int len = strlen(script_cmd) + 11 + 1;
+					if ((ctmp = (char*) malloc(len)) != NULL) {
+						memset(ctmp, 0, len);
+						strcpy(ctmp, "script run ");
+						strcpy(&ctmp[11], script_cmd);
+						script_cmd = ctmp;
+					}
+				}
+				
+				printf("Execute command from commandline: %s\n", script_cmd);
+			}
+		} else {
+			script_cmds_file = argv[argc - 1];
+		}
+	}
+
+	// check command
+	if (executeCommand && (!script_cmd || strlen(script_cmd) == 0)){
+		printf("ERROR: execute command: command not found.\n");
+		return 2;
+	}
+	
+	// set global variables
 	set_my_executable_path();
 	
-	bool usb_present = false;
-	char *script_cmds_file = NULL;
-  
-	sp = uart_open(argv[1]);
+	// open uart
+	if (!waitCOMPort) {
+		sp = uart_open(argv[1]);
+	} else {
+		printf("Waiting for Proxmark to appear on %s ", argv[1]);
+		int openCount = 0;
+		do {
+			sp = uart_open(argv[1]);
+			msleep(1000);
+			printf(".");
+		} while(++openCount < 20 && (sp == INVALID_SERIAL_PORT || sp == CLAIMED_SERIAL_PORT));
+		printf("\n");
+	}
+
+	// check result of uart opening
 	if (sp == INVALID_SERIAL_PORT) {
 		printf("ERROR: invalid serial port\n");
 		usb_present = false;
@@ -254,44 +378,29 @@ int main(int argc, char* argv[]) {
 		usb_present = true;
 		offline = 0;
 	}
-
-	// If the user passed the filename of the 'script' to execute, get it
-	if (argc > 2 && argv[2]) {
-		if (argv[2][0] == 'f' &&  //buzzy, if a word 'flush' passed, flush the output after every log entry.
-			argv[2][1] == 'l' &&
-			argv[2][2] == 'u' &&
-			argv[2][3] == 's' &&
-			argv[2][4] == 'h')
-		{
-			printf("Output will be flushed after every print.\n");
-			flushAfterWrite = 1;
-		}
-		else
-		script_cmds_file = argv[2];
-	}
-
+	
 	// create a mutex to avoid interlacing print commands from our different threads
 	pthread_mutex_init(&print_lock, NULL);
 
 #ifdef HAVE_GUI
 #ifdef _WIN32
-	InitGraphics(argc, argv, script_cmds_file, usb_present);
+	InitGraphics(argc, argv, script_cmds_file, script_cmd, usb_present);
 	MainGraphics();
 #else
 	char* display = getenv("DISPLAY");
 
 	if (display && strlen(display) > 1)
 	{
-		InitGraphics(argc, argv, script_cmds_file, usb_present);
+		InitGraphics(argc, argv, script_cmds_file, script_cmd, usb_present);
 		MainGraphics();
 	}
 	else
 	{
-		main_loop(script_cmds_file, usb_present);
+		main_loop(script_cmds_file, script_cmd, usb_present);
 	}
 #endif
 #else
-	main_loop(script_cmds_file, usb_present);
+	main_loop(script_cmds_file, script_cmd, usb_present);
 #endif	
 
 	// Clean up the port