+++ CHANGELOG.md
@@ -12,8 +12,10 @@ This project uses the changelog in accordance with [keepchangelog](http://keepac
 - Added `hf mf nested` mode: autosearch keys for attack (from well known keys) (Merlok)
 - `hf mf nested` Check keys after they have found (Merlok)
 - `hf mf chk` Move main cycle to arm (Merlok)
+- Changed proxmark command line parameter `flush` to `-f` or `-flush` (Merlok)
 
 ### Fixed
+- Changed start sequence in Qt mode (fix: short commands hangs main Qt thread) (Merlok)
 
 ### Added
 - Added PAC/Stanley detection to lf search (marshmellow)
@@ -27,6 +29,9 @@ This project uses the changelog in accordance with [keepchangelog](http://keepac
 - Added to `hf mf nested` after attack it checks all found keys on non-open sectors (Merlok)
 - `hf mf chk` Added setings to set iso14443a operations timeout. default timeout set to 500us (Merlok)
 - Added to `hf mf nested` parameters `s` and `ss` for checking slow cards (Merlok)
+- Added to proxmark command line parameters `w` - wait 20s for serial port (Merlok)
+- Added to proxmark command line parameters `c` and `l` - execute command and lua script from command line (Merlok)
+- Added to proxmark ability to execute commands from stdin (pipe) (Merlok)
 
 ## [3.0.1][2017-06-08]